LayoutTests/accessibility/presentational-children-expected.txt

This tests that roles with children presentational true report no children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


button childrenCount: 0
checkbox childrenCount: 0
img childrenCount: 0
math childrenCount: 0
menuitemcheckbox childrenCount: 0
menuitemradio childrenCount: 0
option childrenCount: 0
progressbar childrenCount: 0
radio childrenCount: 0
scrollbar childrenCount: 0
separator childrenCount: 0
slider childrenCount: 0
switch childrenCount: 0
tab childrenCount: 0
PASS successfullyParsed is true

TEST COMPLETE
